requirement.txt
absl-py==0.1.10
alembic==0.9.6
amqp==2.2.2
arrow==0.12.0
asn1crypto==0.24.0
attrs==17.3.0
Automat==0.6.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
bleach==1.5.0
blinker==1.4
bs4==0.0.1
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
crypto==1.4.1
cryptography==2.1.4
cssselect==1.0.3
cx-Freeze==5.0.2
cycler==0.10.0
dateparser==0.6.0
decorator==4.1.2
Django==1.11.6
django-debugtools==1.7.2
django-haystack==2.6.1
django-redis==4.8.0
django-request==1.5.3
docopt==0.6.2
dominate==2.3.1
entrypoints==0.2.3
et-xmlfile==1.0.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
html5lib==0.9999999
hyperlink==17.3.1
idna==2.6
image==1.5.16
imageio==2.2.0
incremental==17.5.0
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
jieba==0.39
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.3
kombu==4.1.0
logger==1.4
lxml==3.7.3
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
moment==0.8.2
Naked==0.1.31
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.13.3
olefile==0.44
openpyxl==2.4.9
pandas==0.20.3
pandocfilters==1.4.2
parsel==1.3.1
parso==0.1.1
pickleshare==0.7.4
Pillow==4.3.0
pluggy==0.6.0
prettytable==0.7.2
prompt-toolkit==1.0.15
protobuf==3.5.1
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pygame==1.9.3
Pygments==2.2.0
PyMySQL==0.7.11
pyOpenSSL==17.5.0
pyparsing==2.2.0
PyQt5==5.9
pyqt5-tools==5.9.0.1.2
pytest==3.3.1
python-csv==0.0.10
python-dateutil==2.6.1
python-docx==0.8.6
python-editor==1.0.3
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qrcode==5.3
qtconsole==4.3.1
queuelib==1.4.2
redis==2.10.6
regex==2017.12.12
requests==2.18.4
ruamel.yaml==0.15.35
scikit-learn==0.19.1
scipy==1.0.0
Scrapy==1.4.0
scrapy-redis==0.6.8
selenium==3.7.0
service-identity==17.0.0
shellescape==3.4.1
simplegeneric==0.8.1
sip==4.19.3
six==1.11.0
sklearn==0.0
SQLAlchemy==1.1.15
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
testpath==0.3.1
threadpool==1.3.2
times==0.7
tornado==4.5.2
traitlets==4.3.2
tushare==1.1.3
Twisted==17.9.0
tzlocal==1.5.1
urllib3==1.22
vboxapi==1.0
vine==1.1.4
virtualenv==15.1.0
virtualenvwrapper-win==1.2.4
visitor==0.1.3
w3lib==1.18.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
Whoosh==2.7.4
widgetsnbextension==3.0.8
win32compat==221.26
win32core==221.36
wordcloud==1.3.3
WTForms==2.1
xlrd==1.1.0
xlwt==1.3.0
you-get==0.4.1011
zope.interface==4.4.3
